pyproject.toml
[project]
name = "llm-analysis-quiz-solver"
version = "1.0.0"
description = "Autonomous agent for solving data-related quiz tasks using LLMs"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    
    # Web scraping
    "playwright>=1.48.0",
    "aiohttp>=3.10.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    
    # Data processing (commonly needed)
    "pandas>=2.2.0",
    "numpy>=2.1.0",
    "openpyxl>=3.1.0",
    
    # PDF processing
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.11.0",
    "tabula-py>=2.9.0",
    
    # Image processing
    "Pillow>=10.4.0",
    "opencv-python>=4.10.0",
    
    # Visualization
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "plotly>=5.24.0",
    
    # Audio/Video
    "SpeechRecognition>=3.10.0",
    "moviepy>=1.0.0",
    
    # Network analysis
    "networkx>=3.3",
    
    # Geospatial
    "geopandas>=1.0.0",
    "folium>=0.17.0",
    
    # ML/Statistics
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
    "statsmodels>=0.14.0",
    
    # Text processing
    "nltk>=3.9",
    "textblob>=0.18.0",
    
    # Utilities
    "requests>=2.32.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[project.scripts]
llm-analysis = "main:app"